requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
requests==2.31.0
httpx==0.25.1
python-dotenv==1.0.0
Pillow>=10.0.0
opencv-python-headless==4.8.1.78
gTTS==2.4.0
moviepy==1.0.3
numpy==1.26.2
python-multipart==0.0.6